refactor: '요일:시간' 변환 함수 utils 폴더로 이동
